# frontend/server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Bitmoi service",
        "title": "Bitmoi API",
        "contact": {
            "name": "API Support",
            "email": "bitmoiigson@gmail.com"
        },
        "license": {
            "name": "CC BY-NC-SA 4.0",
            "url": "https://creativecommons.org/licenses/by-nc-sa/4.0/"
        },
        "version": "1.0"
    },
    "host": "api.bitmoi.co.kr",
    "basePath": "/",
    "paths": {
...
        "/reissueAccess": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "access token을 재발급 합니다.",
                "parameters": [
                    {
                        "description": "refresh token",
                        "name": "ReissueAccessTokenRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ReissueAccessTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReissueAccessTokenResponse"
                        }
                    }
                }
            }
        },
...
        "/user": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "사용자를 생성합니다.",
                "parameters": [
                    {
                        "description": "유저 생성 요청에 필요한 정보",
                        "name": "CreateUserRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/user/address": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "사용자의 Metamask 계좌를 업데이트 합니다.",
                "parameters": [
                    {
                        "description": "Metamask 주소",
                        "name": "LoginUserRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.MetamaskAddressRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/user/checkId": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "ID 중복 여부를 확인합니다.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "체크할 ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/user/checkNickname": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "닉네임 중복 여부를 확인합니다.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "체크할 닉네임",
                        "name": "nickname",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "사용자의 ID와 PW를 확인하여 토큰을 발급합니다.",
                "parameters": [
                    {
                        "description": "ID와 PW",
                        "name": "LoginUserRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.LoginUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.LoginUserResponse"
                        }
                    }
                }
            }
        },
        "/user/profile": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "사용자의 프로필 이미지를 업데이트 합니다.",
                "parameters": [
                    {
                        "type": "file",
                        "description": "profile image",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/verifyToken": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "발급했던 access token을 검증합니다.",
                "parameters": [
                    {
                        "description": "access token",
                        "name": "VerifyTokenRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.VerifyTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.UserResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "api.CreateUserRequest": {
            "type": "object",
            "required": [
                "email",
                "nickname",
                "password",
                "user_id"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string",
                    "maxLength": 10,
                    "minLength": 1
                },
                "oauth_uid": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 8
                },
                "user_id": {
                    "type": "string",
                    "maxLength": 15,
                    "minLength": 5
                }
            }
        },
...
        "api.LoginUserRequest": {
            "type": "object",
            "required": [
                "password",
                "user_id"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "minLength": 8
                },
                "user_id": {
                    "type": "string",
                    "maxLength": 15,
                    "minLength": 5
                }
            }
        },
        "api.LoginUserResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "access_token_expires_at": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                },
                "refresh_token_expires_at": {
                    "type": "string"
                },
                "session_id": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/api.UserResponse"
                }
            }
        },
        "api.MetamaskAddressRequest": {
            "type": "object",
            "required": [
                "addr"
            ],
            "properties": {
                "addr": {
                    "type": "string"
                }
            }
        },
...
        "api.ReissueAccessTokenRequest": {
            "type": "object",
            "required": [
                "refresh_token"
            ],
            "properties": {
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "api.ReissueAccessTokenResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "access_token_expires_at": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/api.UserResponse"
                }
            }
        },
...
        "api.UserResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "metamask_address": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "password_changed_at": {
                    "type": "string"
                },
                "photo_url": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "api.VerifyTokenRequest": {
            "type": "object",
            "required": [
                "token"
            ],
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
...
    }
}